+++ Exceptions.cs
@@ -15,13 +15,19 @@ namespace alphappy.TAMacro
             public TAMacroException(string message) : base(message) { }
             public TAMacroException(string message, Exception innerException) : base(message, innerException) { }
             protected TAMacroException(SerializationInfo info, StreamingContext context) : base(info, context) { }
+            public string MultiLineMessage => Message.Replace("  ", "\n ");
+        }
+        public class MacroRuntimeException : TAMacroException
+        {
+            public MacroRuntimeException(Macro macro, string message) : base($"An exception occured while running {macro.FullName}:  {message}") { }
         }
         public class IllegalCommandException : TAMacroException
         {
             public IllegalCommandException() { }
-            public IllegalCommandException(string message) : base(message) { }
-            public IllegalCommandException(string message, Exception innerException) : base(message, innerException) { }
-            protected IllegalCommandException(SerializationInfo info, StreamingContext context) : base(info, context) { }
+        }
+        public class InvalidExecuteTargetException : MacroRuntimeException
+        {
+            public InvalidExecuteTargetException(Macro macro, string message) : base(macro, $"`>execute` command failed:  {message}") { }
         }
     }
 }