+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-### TAMacro 0.8.4.3 (preview)
+### TAMacro 0.8.4.4 (preview)
 - **BREAKING CHANGE:** `!execute` now accepts Unix-style relative paths.
   - Existing `!execute` commands which were absolute paths in prior versions will be interpreted as relative paths in this version.
   - Existing absolute paths must now start with `/` (e.g., `>execute foo/bar/baz` must be changed to `>execute /foo/bar/baz`) or be converted to relative paths.
 - Display panel now indicates whether recording is in progress.
+- Exceptions rewritten a bit.
 - Remove leftover debug logging (hopefully for the last time).
 
 ### TAMacro 0.8.3.29